add local songs in home screen and fix navigation issues

+++ app/src/main/java/com/mobile/felix/musicapp/core/data/local/entity/SongEntity.kt
@@ -8,9 +8,9 @@ import androidx.room.PrimaryKey
 )
 data class SongEntity(
     @PrimaryKey(autoGenerate = false)
-    val trackId: Int,
+    val trackId: Long,
     val collectionId: Long,
-    val artistId: Int,
+    val artistId: Long,
     val wrapperType: String,
     val kind: String,
     val artistName: String,
