+++ app/src/main/java/com/mobile/felix/musicapp/core/domain/Song.kt
@@ -2,9 +2,9 @@ package com.mobile.felix.musicapp.core.domain
 
 
 data class Song(
-    val trackId: Int?,
+    val trackId: Long?,
     val collectionId: Long?,
-    val artistId: Int?,
+    val artistId: Long?,
     val wrapperType: String?,
     val kind: String?,
     val artistName: String?,
