+++ app/src/main/java/com/mobile/felix/musicapp/feature/song/domain/source/SongLocalDataSource.kt
@@ -3,5 +3,5 @@ package com.mobile.felix.musicapp.feature.song.domain.source
 import com.mobile.felix.musicapp.core.domain.Song
 
 interface SongLocalDataSource {
-    suspend fun getSong(id: Int) : Song?
+    suspend fun getSong(id: Long) : Song?
 }
